[Debug] Solve the import problem + Renamed main.py
Solved the import problem that required the setup.sh to be sourced
in the first place. The import system works properly if all the
entrypoint files are on the project root, so I moved the main.py
from root/appSiamFC to root and renamed it vis_app.py (visualization
application). Also solved a bug on the call of get_annotations in
appSiamFC/app_utils.py